services/audit-service/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.27.0
pydantic>=2.10.0
pydantic-settings>=2.7.0
motor>=3.3.2
kafka-python>=2.0.2
prometheus-client>=0.19.0
python-dotenv>=1.0.0
